VAdmin/pyproject.toml
[tool.poetry]
name = "vadmin"
version = "0.1.0"
description = ""
authors = ["DevJsmJsr <jsmajsr@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.35.19"
Django = "^5.1.1"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
django-rest-polymorphic = "^0.1.10"
djangorestframework-datatables = "^0.7.2"
psycopg2-binary = "^2.9.9"
zappa = "^0.59.0"
django-rest-passwordreset = "^1.4.1"
djangorestframework-simplejwt = "^5.3.1"
coreapi = "^2.3.3"
botocore = "^1.35.19"
setuptools = "74.1.2"

[tool.poetry.group.dev.dependencies]
python = "^3.11"
boto3 = "^1.35.19"
Django = "^5.1.1"
django-cors-headers = "^4.4.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
django-rest-polymorphic = "^0.1.10"
djangorestframework-datatables = "^0.7.2"
psycopg2-binary = "^2.9.9"
zappa = "^0.59.0"
djangorestframework-simplejwt = "^5.3.1"
coreapi = "^2.3.3"
botocore = "^1.35.19"
setuptools = "74.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
